+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,6 +1,20 @@
 package com.example.example
 
 import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.engine.FlutterEngine
+import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugins.GeneratedPluginRegistrant
 
-class MainActivity: FlutterActivity() {
+class MainActivity : FlutterActivity() {
+    private val openIntent: String = "openIntentChannel"
+
+    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
+        GeneratedPluginRegistrant.registerWith(flutterEngine)
+
+        MethodChannel(flutterEngine.dartExecutor, openIntent).setMethodCallHandler(
+            IndentHandler(
+                activity
+            )
+        )
+    }
 }